+++ lib/src/flowr/mixin.dart
@@ -1,3 +1,4 @@
+export 'package:flowr/src/flowr/mixin/dispose.dart';
 export 'package:flowr/src/flowr/mixin/auto_dispose.dart';
 export 'package:flowr/src/flowr/mixin/loggable.dart';
 export 'package:flowr/src/flowr/mixin/updatable.dart';