+++ lib/src/flowr/mixin/updatable.dart
@@ -3,35 +3,55 @@ import 'package:flowr/src/flowr/base.dart';
 import 'package:flowr/src/flowr/mixin/slowly.dart';
 
 /// 添加[update]方法, 自动捕获异常
-mixin UpdatableMx<T> on BaseFlowR<T>, SlowlyMx<T> {
+mixin UpdatableMx<T> on BaseFlowR<T>, SlowlyMx {
   /// 执行一个异步操作, 并更新状态
   /// 不建议对本方法进行二次包装, 因此返回值强制为 void
   Future<void> update(
     FutureOr<T> Function(T old) update, {
     Function(Object e, StackTrace s)? onError,
+    int slowlyMs = 100,
+    Object? debounceTag,
+    Object? throttleTag,
   }) async =>
-      await updateRaw((old) => update(old), onError: onError);
+      await updateRaw(
+        (old) => update(old),
+        onError: onError,
+        slowlyMs: slowlyMs,
+        debounceTag: debounceTag,
+        throttleTag: throttleTag,
+      );
 
   /// for advance user
   ///   you can sync update value, and get the return value
+  /// [slowlyMs] <=0, will ignore debounce
   FutureOr<T?> updateRaw(
     FutureOr<T> Function(T old) update, {
     Function(Object e, StackTrace s)? onError,
-    int? debounceMs,
-    Object? slowlyTag,
+    int slowlyMs = 100,
+    Object? debounceTag,
+    Object? throttleTag,
   }) async {
     try {
-      /// debounce
-      assert(debounceMs == null || slowlyTag != null,
-          'slowlyKey can not be null when debounceMs is set');
-      if (debounceMs != null && slowlyTag != null) {
-        final deFunc = await slowly.debounce(
-          slowlyTag,
-          update,
-          duration: Duration(milliseconds: debounceMs),
-        );
-        if (deFunc == null) return null;
-        update = deFunc;
+      if (slowlyMs > 0) {
+        if (debounceTag != null) {
+          /// debounce
+          final deFunc = await slowly.debounce(
+            debounceTag,
+            update,
+            duration: Duration(milliseconds: slowlyMs),
+          );
+          if (deFunc == null) return null;
+          update = deFunc;
+        } else if (throttleTag != null) {
+          /// throttle
+          final thFunc = slowly.throttle(
+            throttleTag,
+            update,
+            duration: Duration(milliseconds: slowlyMs),
+          );
+          if (thFunc == null) return null;
+          update = thFunc;
+        }
       }
 
       final data = update(value);